requirements.txt
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
boto3==1.28.33
botocore==1.31.33
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
colander==2.0
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.4
django-ckeditor==6.7.0
django-ckeditor-5==0.2.9
django-credit-cards==0.4.1
django-crispy-forms==2.0
django-embed-video==1.4.9
django-js-asset==2.1.0
django-qr-code==3.1.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
dropbox==11.36.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
iso8601==2.0.0
jmespath==1.0.1
lxml==4.9.3
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.0
ply==3.11
protobuf==4.24.1
psycopg2-binary==2.9.7
Py-Authorize==1.4.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.15.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
segno==1.5.2
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
svglib==1.5.1
tinycss2==1.2.1
translationstring==1.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
